requirements.txt
groq>=0.9.0
pydantic>=2.11.7,<3.0.0
python-dotenv==1.0.1
rich==13.7.1
requests==2.31.0
tzdata==2024.1
